Move CellInstance to cell.h
