+++ include/openmc/tallies/filter_cell_instance.h
@@ -2,42 +2,13 @@
 #define OPENMC_TALLIES_FILTER_CELL_INSTANCE_H
 
 #include <cstdint>
-#include <functional> // for hash
 #include <unordered_map>
 #include <vector>
 
 #include <gsl/gsl>
 
+#include "openmc/cell.h"
 #include "openmc/tallies/filter.h"
-
-//==============================================================================
-//! Define an instance of a particular cell
-//==============================================================================
-
-namespace openmc {
-struct CellInstance {
-  //! Check for equality
-  bool operator==(const CellInstance& other) const
-  { return index_cell == other.index_cell && instance == other.instance; }
-
-  gsl::index index_cell;
-  gsl::index instance;
-};
-} // namespace openmc
-
-namespace std {
-template<>
-struct hash<openmc::CellInstance> {
-  // Taken from https://stackoverflow.com/a/17017281
-  std::size_t operator()(const openmc::CellInstance& k) const
-  {
-    std::size_t res = 17;
-    res = 31 * res + std::hash<gsl::index>()(k.index_cell);
-    res = 31 * res + std::hash<gsl::index>()(k.instance);
-    return res;
-  }
-};
-} // namespace std
 
 
 namespace openmc {
